Made changes to prepare for the submission to AI CUP
Added report.tex that contains the report of this project. It still
needs a little more content to be ready.

Deleted Python scripts that are no longer used.

Modified .gitignore to ignore LaTeX temporary files.

Modified README to include the up-to-date help message of palapapa.py.

Added the images directory for the images used in report.tex.

Fixed a type error in palapapa.py.